Migrate to Tauri v2
