+++ rust/Cargo.toml
@@ -14,11 +14,15 @@ readme = "README.md"
 rust-version = "1.76.0"
 version = "0.1.3"
 
+[lib]
+name = "app_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
 prost-build = "0.11.0"
-tauri-build = { version = "1.5.1", features = [] }
+tauri-build = { version = "2.0.0-rc", features = [] }
 
 [dependencies]
 anyhow = "1.0.69"
@@ -37,11 +41,10 @@ serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.115"
 sha2 = "0.10.6"
 sync-unsafe-cell = "0.1.0"
-tauri = { version = "1.6.1", features = [
-    "shell-open",
-    "process-exit",
-    "dialog-open",
-] }
+tauri = { version = "2.0.0-rc", features = [] }
+tauri-plugin-dialog = "2.0.0-rc"
+tauri-plugin-process = "2.0.0-rc"
+tauri-plugin-shell = "2.0.0-rc"
 tempfile = "3.6.0"
 zip = { version = "0.6.6", features = [
     "deflate",
